added basic skel for filter and abstract service
