+++ src/main/java/com/xjeffrose/xio/Filter.java
@@ -1,0 +1,14 @@
+package com.xjeffrose.xio;
+
+abstract class Filter {
+
+  public HttpRequest req;
+
+  Filter() {
+    this.req = req;
+  }
+
+  public void filter() {
+  }
+
+}
