+++ src/main/java/com/xjeffrose/xio/IpFilter.java
@@ -2,7 +2,7 @@ package com.xjeffrose.xio;
 
 import java.net.*;
 
-class IpFilter {
+class IpFilter extends Filter {
 
   IpFilter() {
   }
